Added the methods for getting the truck loads for a given departure date and for updating the truck load and it's load items. Also made some other minor changes

+++ src/main/java/rs/ac/bg/fon/njt/DiploDispatch/truck/TruckController.java
@@ -19,7 +19,7 @@ public class TruckController {
     }
 
     @GetMapping("/manufacturer/{id}")
-    public List<Truck> getTrucksByManufacturer(@PathVariable int id) {
+    public List<Truck> getTrucksByManufacturer(@PathVariable Long id) {
         return truckService.getTrucksByManufacturerId(id);
     }
